fix message limits , added enter to send and auto clear message bar , refactor socket setup into cibtroller and service for readability
- added functionality to send on enter and auto clear message on send

- added context message type for easier message history filtering and limite ddocument sizes and limited history to recent history to ensure the usage is under the token limits

- refactored socketSetup to socketMessageController and AiService
