+++ limedaichatFrontend/src/components/chat/MessageBar.tsx
@@ -40,7 +40,9 @@ const MessageBar: React.FC = () => {
   const addEmoji = (emoji: EmojiClickData) => {
     setMessage((Message) => Message + emoji.emoji);
   };
+
   const sendMessage = () => {
+    if (!Message.trim()) return;
     console.log("sending message");
     console.log(chatType);
 
@@ -56,7 +58,17 @@ const MessageBar: React.FC = () => {
         messageType: "text",
       });
     }
+
+    setMessage("");
   };
+
+  const handleKeyDown = (event: React.KeyboardEvent<HTMLInputElement>) => {
+    if (event.key === "Enter") {
+      event.preventDefault();
+      sendMessage();
+    }
+  };
+
   return (
     <div className="h-[10vh] bg-[#1c1d25] flex justify-center items-center px-8 mb-6 gap-6">
       <div className="flex-1 flex bg-[#2a2b33] rounded-md items-center gap-5 pr-5">
@@ -67,6 +79,7 @@ const MessageBar: React.FC = () => {
           placeholder="输入您的留言"
           value={Message}
           onChange={(e) => setMessage(e.target.value)}
+          onKeyDown={handleKeyDown} // Add this line
         />
         {/* <button className="text-neutral-500 focus:border-none focus:outline-none focus:text-white duration-300 transition-all">
           <GrAttachment className="text-2xl" />
@@ -78,14 +91,16 @@ const MessageBar: React.FC = () => {
           >
             <RiEmojiStickerLine className="text-2xl" />
           </button>
-          <div className="absolute bottom-16 right-0" ref={emojiRef}>
-            <EmojiPicker
-              theme={"dark" as Theme}
-              open={emojiOpen}
-              onEmojiClick={addEmoji}
-              autoFocusSearch={false}
-            />
-          </div>
+          {emojiOpen && (
+            <div className="absolute bottom-16 right-0" ref={emojiRef}>
+              <EmojiPicker
+                theme={"dark" as Theme}
+                open={emojiOpen}
+                onEmojiClick={addEmoji}
+                autoFocusSearch={false}
+              />
+            </div>
+          )}
         </div>
       </div>
       <div>
