+++ limedaichatserver/src/utils/socketSetup.ts
@@ -1,22 +1,8 @@
 import { Socket, Server as SocketIOServer } from "socket.io";
 import { Server } from "http";
-import { UserSocketMap, IMessage } from "@/types";
+import { UserSocketMap } from "@/types";
 import dotenv from "dotenv";
-import Message from "@/models/MessageSchema";
-import Textbook from "@/models/Textbookmodel";
-import { FaissStore } from "@langchain/community/vectorstores/faiss";
-import { OpenAIEmbeddings } from "@langchain/openai";
-import { ChatOpenAI } from "@langchain/openai";
-import { createHistoryAwareRetriever } from "langchain/chains/history_aware_retriever";
-import { createRetrievalChain } from "langchain/chains/retrieval";
-import { createStuffDocumentsChain } from "langchain/chains/combine_documents";
-import {
-  ChatPromptTemplate,
-  MessagesPlaceholder,
-} from "@langchain/core/prompts";
-import { BaseMessage, HumanMessage, AIMessage } from "@langchain/core/messages";
-import { Document } from "langchain/document";
-import { cleanText, splitTextIntoChunks } from "@/utils/textUtils";
+import { handleSendMessage } from "@/controllers/SocketMessageController";
 
 dotenv.config();
 
@@ -39,208 +25,6 @@ const setupSocket = (server: Server): SocketIOServer => {
 
   const userSocketMap: UserSocketMap = new Map();
 
-  const handleSendMessage = async (message: Partial<IMessage>) => {
-    console.log("Received message: ", message);
-    const { sender, content, receiver, receiverModel, isAI, messageType } =
-      message;
-
-    try {
-      // Fetch the selected textbook based on the receiver ID
-      console.log("Receiver ID: ", receiver);
-
-      const textbook = await Textbook.findById(receiver);
-      if (!textbook) {
-        throw new Error("Textbook not found");
-      }
-
-      // Create and save the user message
-      const createdMessage = await Message.create({
-        sender,
-        receiver: textbook._id,
-        receiverModel,
-        senderModel: "User",
-        isAI,
-        messageType,
-        content: content,
-        timeStamp: new Date(),
-      });
-
-      // Send the user message back to the user (if needed)
-      const userSocketId = userSocketMap.get(sender?.toString() || "");
-      if (userSocketId) {
-        io.to(userSocketId).emit("receive-message", createdMessage);
-      }
-
-      // Load the vector store for the textbook
-      const vectorStore = await FaissStore.loadFromPython(
-        textbook.vectorStorePath as string,
-        new OpenAIEmbeddings()
-      );
-      console.log("Vector store loaded");
-
-      // Retrieve the chat history between the user and the textbook
-      const chatHistoryMessages = await Message.find({
-        $or: [
-          { sender: sender, receiver: textbook._id },
-          { sender: textbook._id, receiver: sender },
-        ],
-      }).sort({ timeStamp: 1 });
-      console.log("Chat history loaded");
-
-      // Convert chat history to BaseMessages
-      const chat_history: BaseMessage[] = chatHistoryMessages.map((msg) => {
-        if (msg.sender.toString() === sender?.toString()) {
-          return new HumanMessage(msg.content as string);
-        } else {
-          return new AIMessage(msg.content as string);
-        }
-      });
-      console.log("Chat history converted");
-
-      // Set up the OpenAI model
-      const llm = new ChatOpenAI({
-        model: "gpt-4o",
-        openAIApiKey: process.env.OPENAI_API_KEY,
-        temperature: 0.2,
-      });
-      console.log("OpenAI model set up");
-
-      // Contextualize question prompt
-      const contextualizeQSystemPrompt = `
-鉴于聊天记录和用户的最新提问（可能引用了聊天记录中的内容），
-请用中文重新表述一个独立的问题，使其在没有聊天记录的情况下也能被理解。
-请不要回答问题，只需在需要时重新表述，否则按原样返回。`;
-
-      const contextualizeQPrompt = ChatPromptTemplate.fromMessages([
-        { role: "system", content: contextualizeQSystemPrompt },
-        new MessagesPlaceholder("chat_history"),
-        { role: "user", content: "{input}" },
-      ]);
-      console.log("Contextualize question prompt set up");
-
-      // Create a history-aware retriever
-      const retriever = vectorStore.asRetriever();
-      const historyAwareRetriever = await createHistoryAwareRetriever({
-        llm,
-        retriever,
-        rephrasePrompt: contextualizeQPrompt,
-      });
-      console.log("History-aware retriever set up");
-
-      // Answer question prompt
-      const qaSystemPrompt = `
-你是一个负责问答任务的助手。请使用以下检索到的上下文来回答问题。
-你的回答应该使用中文。
-如果你不知道答案，只需说“我不知道”。
-请尽量简洁，用不超过三句话来回答。
-
-{context}`;
-      console.log("Answer question prompt set up");
-
-      const qaPrompt = ChatPromptTemplate.fromMessages([
-        { role: "system", content: qaSystemPrompt },
-        new MessagesPlaceholder("chat_history"),
-        { role: "user", content: "{input}" },
-      ]);
-      console.log("QA prompt set up");
-
-      // Create the question-answering chain
-      const questionAnswerChain = await createStuffDocumentsChain({
-        llm,
-        prompt: qaPrompt,
-      });
-      console.log("QA chain set up");
-
-      // Create the retrieval chain
-      const ragChain = await createRetrievalChain({
-        retriever: historyAwareRetriever,
-        combineDocsChain: questionAnswerChain,
-      });
-      console.log("Retrieval chain set up");
-      console.log(chat_history);
-      console.log(content);
-
-      // Generate the AI response
-      const response = await ragChain.invoke({
-        input: content as string,
-        chat_history,
-      });
-      console.log("AI response generated");
-      console.log("AI response: ", response);
-
-      const aiMessage = await Message.create({
-        sender: textbook._id,
-        receiver: sender,
-        senderModel: "Textbook",
-        receiverModel: "User",
-        isAI: true,
-        messageType: "text",
-        content: response.answer,
-        timeStamp: new Date(),
-      });
-      console.log("AI message created");
-
-      if (userSocketId) {
-        io.to(userSocketId).emit("receive-message", aiMessage);
-      }
-
-      console.log("Response context: ", response.context);
-
-      const contextDocuments = response.context as Document[];
-
-      if (Array.isArray(contextDocuments) && contextDocuments.length > 0) {
-        const selectedDocuments = contextDocuments.slice();
-
-        for (let i = 0; i < selectedDocuments.length; i++) {
-          const doc = selectedDocuments[i];
-          console.log("Original Document Content:", doc.pageContent);
-
-          const cleanedPageContent = cleanText(doc.pageContent);
-
-          console.log("Cleaned Document Content:", cleanedPageContent);
-
-          const messageHeader = `【片段${i + 1}】\n`;
-
-          const maxMessageLength = 1000;
-
-          const contentChunks = splitTextIntoChunks(
-            cleanedPageContent,
-            maxMessageLength - messageHeader.length
-          );
-
-          let isFirstChunk = true;
-
-          for (let j = 0; j < contentChunks.length; j++) {
-            const chunkContent =
-              (isFirstChunk ? messageHeader : "") + contentChunks[j];
-            isFirstChunk = false;
-
-            const contextMessage = await Message.create({
-              sender: textbook._id,
-              receiver: sender,
-              senderModel: "Textbook",
-              receiverModel: "User",
-              isAI: true,
-              messageType: "text",
-              content: chunkContent,
-              timeStamp: new Date(),
-            });
-            console.log("Context message created");
-            console.log("Context message: ", contextMessage);
-
-            if (userSocketId) {
-              io.to(userSocketId).emit("receive-message", contextMessage);
-            }
-          }
-        }
-      } else {
-        console.log("No context documents available.");
-      }
-    } catch (error) {
-      console.error("Error handling message:", error);
-    }
-  };
-
   const disconnect = (socket: Socket) => {
     console.log("disconnected", socket.id);
     for (const [userId, socketId] of userSocketMap.entries()) {
@@ -261,7 +45,9 @@ const setupSocket = (server: Server): SocketIOServer => {
       console.log("User ID not provided during connection.");
     }
 
-    socket.on("send-message", handleSendMessage);
+    socket.on("send-message", (message) =>
+      handleSendMessage(io, userSocketMap, message)
+    );
     socket.on("disconnect", () => disconnect(socket));
   });
